Updated cases of BigInteger to Long and updated BayouDB to work with WriteID and BayouData.
git-svn-id: file:///home/sheu/svn/bayou@20 f546f330-0b42-417b-95bf-8a63a2b938e6

+++ ServerID.java
@@ -1,11 +1,11 @@
-import java.math.BigInteger;
+
 
 public class ServerID implements Comparable<ServerID>
 {
 	private ServerID sid;
-	private BigInteger acceptStamp;
+	private Long acceptStamp;
 
-	public ServerID( ServerID sid, BigInteger acceptStamp )
+	public ServerID( ServerID sid, Long acceptStamp )
 	{
 		this.sid = sid;
 		this.acceptStamp = acceptStamp;
